last commit kaj kore deploy hoy nai , oita ektu fix korsi

+++ src/main/java/com/jarvisatt/attendance/dto/SessionDtos.java
@@ -50,6 +50,6 @@ public final class SessionDtos {
             OffsetDateTime startedAt,
             OffsetDateTime endedAt,
             Double radiusMeters,
-            int attendanceCount
+            long attendanceCount
     ) {}
 }
